splitting off fco (brightlite) metrics from general metrics
